delete .xprojes
