+++ tests/src/ApplicationException.cs
@@ -1,10 +1,11 @@
-﻿#region Apache License
+#if NETCOREAPP1_0
+#region Apache License
 //
-// Licensed to the Apache Software Foundation (ASF) under one or more
+// Licensed to the Apache Software Foundation (ASF) under one or more 
 // contributor license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright ownership.
+// this work for additional information regarding copyright ownership. 
 // The ASF licenses this file to you under the Apache License, Version 2.0
-// (the "License"); you may not use this file except in compliance with
+// (the "License"); you may not use this file except in compliance with 
 // the License. You may obtain a copy of the License at
 //
 // http://www.apache.org/licenses/LICENSE-2.0
@@ -25,3 +26,4 @@ namespace log4net.Tests
 	{
 	}
 }
+#endif
